+++ uaclib/UsbDevice.cpp
@@ -97,8 +97,10 @@ bool USBDevice::ParseDescriptors(BYTE *configDescr, DWORD length)
 	return TRUE;
 }
 
+//bool USBDevice::SendUsbControl(int dir, int type, int recipient, int request, int value, int index,
+//				   unsigned char *buff, int size, ULONG *lengthTransferred)
 bool USBDevice::SendUsbControl(int dir, int type, int recipient, int request, int value, int index,
-				   unsigned char *buff, int size, ULONG *lengthTransferred)
+				   unsigned char *buff, int size, UINT *lengthTransferred)
 {
 	bool retVal = FALSE;
 	WINUSB_SETUP_PACKET packet;
@@ -131,7 +133,8 @@ KUSB_HANDLE USBDevice::FindDevice()
 	KLST_DEVINFO_HANDLE DeviceInfo = NULL;
 	KLST_DEVINFO_HANDLE tmpDeviceInfo = NULL;
 
-	ULONG deviceCount = 0;
+//	ULONG deviceCount = 0;
+	UINT deviceCount = 0;
 	m_errorCode = ERROR_SUCCESS;
 
 	// Get the device list
@@ -199,7 +202,8 @@ KUSB_HANDLE USBDevice::FindDevice()
 bool USBDevice::InitDevice()
 {
 	BYTE configDescriptorBuffer[4096];
-	ULONG lengthTransferred;
+//	ULONG lengthTransferred;
+	UINT lengthTransferred;
 
 	m_deviceMutex = CreateMutex(NULL, FALSE, "Global\\ASIOUAC2");
 	if(GetLastError() == ERROR_ALREADY_EXISTS)
